requirements.txt
astroid==2.2.2
cloudpickle==0.8.0
cycler==0.10.0
dask==1.1.3
decorator==4.3.2
freeze==1.0.10
imutils==0.5.2
isort==4.3.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.3
mccabe==0.6.1
networkx==2.2
numpy==1.16.2
opencv-contrib-python==4.0.0.21
pandas==0.24.1
Pillow==5.4.1
psutil==5.5.1
pylint==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
toolz==0.9.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.31.1
typed-ast==1.2.0
wrapt==1.11.1
